chore: Clean unused blobs
There's a lot of PSDs in Pizzacus/satania.moe, which isn't needed for
suisei-cn/hosimati.suisei.moe (and it's actually unlikely that we will
have a directory like this as well). Therefore, the history of "/High
Resolution and PSDs" is cleaned from this repository by
bfg-repo-cleaner for the experience of cloning and deploying.

To see the original tree, see branch "satania-master".